+++ common-jpa/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.firstlogistics.common.jpa.config.JpaAuditingConfig
+common.jpa.config.JpaAuditingConfig
