+++ sample-service/src/main/java/com/firstlogistics/sampleservice/application/ItemCommandService.java
@@ -1,11 +1,11 @@
-package com.firstlogistics.sampleservice.application.item;
+package com.firstlogistics.sampleservice.application;
 
-import com.firstlogistics.sampleservice.application.item.command.CreateItemCommand;
-import com.firstlogistics.sampleservice.application.item.command.UpdateItemCommand;
-import com.firstlogistics.sampleservice.domain.item.Item;
-import com.firstlogistics.sampleservice.domain.item.ItemRepository;
-import com.firstlogistics.sampleservice.domain.item.exception.ItemErrorCode;
-import com.firstlogistics.sampleservice.domain.item.exception.ItemException;
+import com.firstlogistics.sampleservice.application.command.CreateItemCommand;
+import com.firstlogistics.sampleservice.application.command.UpdateItemCommand;
+import com.firstlogistics.sampleservice.infrastructure.entity.Item;
+import com.firstlogistics.sampleservice.domain.repository.ItemRepository;
+import com.firstlogistics.sampleservice.domain.exception.ItemErrorCode;
+import com.firstlogistics.sampleservice.domain.exception.ItemException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
