+++ sample-service/src/main/java/com/firstlogistics/sampleservice/application/ItemQueryService.java
@@ -1,10 +1,10 @@
-package com.firstlogistics.sampleservice.application.item;
+package com.firstlogistics.sampleservice.application;
 
-import com.firstlogistics.sampleservice.domain.item.Item;
-import com.firstlogistics.sampleservice.domain.item.ItemRepository;
-import com.firstlogistics.sampleservice.domain.item.exception.ItemErrorCode;
-import com.firstlogistics.sampleservice.domain.item.exception.ItemException;
-import com.firstlogistics.sampleservice.presentation.item.dto.ItemResponse;
+import com.firstlogistics.sampleservice.infrastructure.entity.Item;
+import com.firstlogistics.sampleservice.domain.repository.ItemRepository;
+import com.firstlogistics.sampleservice.domain.exception.ItemErrorCode;
+import com.firstlogistics.sampleservice.domain.exception.ItemException;
+import com.firstlogistics.sampleservice.presentation.dto.ItemResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
