+++ sample-service/src/main/java/com/firstlogistics/sampleservice/application/command/CreateItemCommand.java
@@ -1,4 +1,4 @@
-package com.firstlogistics.sampleservice.application.item.command;
+package com.firstlogistics.sampleservice.application.command;
 
 public record CreateItemCommand(String name, String description) {
 }
