+++ sample-service/src/main/java/com/firstlogistics/sampleservice/application/command/UpdateItemCommand.java
@@ -1,4 +1,4 @@
-package com.firstlogistics.sampleservice.application.item.command;
+package com.firstlogistics.sampleservice.application.command;
 
 public record UpdateItemCommand(Long id, String name, String description) {
 }
