+++ sample-service/src/main/java/com/firstlogistics/sampleservice/domain/exception/ItemErrorCode.java
@@ -1,4 +1,4 @@
-package com.firstlogistics.sampleservice.domain.item.exception;
+package com.firstlogistics.sampleservice.domain.exception;
 
 import common.response.ErrorCode;
 import lombok.Getter;
