--- sample-service/src/main/java/com/firstlogistics/sampleservice/domain/item/Item.java
@@ -1,47 +1,0 @@
-package com.firstlogistics.sampleservice.domain.item;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "items")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EntityListeners(AuditingEntityListener.class)
-public class Item {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    private String description;
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    public static Item create(String name, String description) {
-        Item item = new Item();
-        item.name = name;
-        item.description = description;
-        return item;
-    }
-
-    public void update(String name, String description) {
-        this.name = name;
-        this.description = description;
-    }
-}
