+++ sample-service/src/main/java/com/firstlogistics/sampleservice/domain/repository/ItemRepository.java
@@ -1,4 +1,6 @@
-package com.firstlogistics.sampleservice.domain.item;
+package com.firstlogistics.sampleservice.domain.repository;
+
+import com.firstlogistics.sampleservice.infrastructure.entity.Item;
 
 import java.util.List;
 import java.util.Optional;
