--- sample-service/src/main/java/com/firstlogistics/sampleservice/infrastructure/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.firstlogistics.sampleservice.infrastructure.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
