+++ sample-service/src/main/java/com/firstlogistics/sampleservice/infrastructure/item/ItemJpaStore.java
@@ -1,6 +1,6 @@
 package com.firstlogistics.sampleservice.infrastructure.item;
 
-import com.firstlogistics.sampleservice.domain.item.Item;
+import com.firstlogistics.sampleservice.infrastructure.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemJpaStore extends JpaRepository<Item, Long> {
