+++ sample-service/src/test/java/com/firstlogistics/sampleservice/SampleServiceApplicationTests.java
@@ -2,8 +2,10 @@ package com.firstlogistics.sampleservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class SampleServiceApplicationTests {
 
 	@Test
